fix tests and environments
