+++ .env
@@ -4,12 +4,11 @@ POSTGRES_PASSWORD=polkastats
 POSTGRES_DATABASE=polkastats
 POSTGRES_HOST=localhost
 POSTGRES_PORT=5432
+WS_PROVIDER_URL='ws://host.docker.internal:9944'
+TYPE_PROVIDER=quartz
 
 #Environment variables for tests
-TESTS_WS_PROVIDER_URL='ws://host.docker.internal:9944'
-TESTS_UNIQUE_WS_ENDPOINT=wss://ws-quartz-dev.unique.network
-TESTS_TYPE_PROVIDER=quartz
-TESTS_GRAPHQL_URL='http://localhost:3031/graphql'
-TESTS_CHAIN_VERSION=918001
+TESTS_UNIQUE_WS_ENDPOINT=ws://127.0.0.1:9944
+TESTS_GRAPHQL_URL='http://localhost:3031/v1/graphql'
 TESTS_CRAWLER_API_REPOSITORY=https://github.com/UniqueNetwork/unique-explorer-api/
 TESTS_CRAWLER_BRANCH=v1.6.3
