+++ e2e/scripts/prepare-scan-api.sh
