+++ e2e/utils/retry.ts
@@ -1,8 +1,12 @@
-export const retry = async (callback: () => Promise<any>, timeout = 3000) => {
+export const retry = async (
+  callback: () => Promise<any>,
+  sleepTime = 500,
+  timeout = 3000,
+) => {
   const now = Date.now();
   const repeat = async () => {
     try {
-      return await callback();
+      return await setTimeout(await callback, sleepTime);
     } catch (error) {
       if (Date.now() - now < timeout) {
         await repeat();
