+++ package.json
@@ -19,10 +19,11 @@
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
     "test:e2e": "jest --config ./apps/web-api/test/jest-e2e.json",
-    "tests:run-chain": "(cd tests/scripts ; ./run-chain.sh ../../.env)",
-    "tests:prepare-scan-api": "(cd tests/scripts ; ./prepare-scan-api.sh ../../.env)",
-    "tests:run-scan-api": "(cd tests/scripts ; ./run-scan-api.sh ../../.env)",
-    "tests:run-environment": "npm run tests:run-chain ; npm run tests:run-scan-api",
+    "tests": "mocha --timeout 500000 --exit -r ts-node/register './e2e/test/*.test.ts'",
+    "tests:chain-start": "docker run -p 9933:9933 -p 9944:9944 --name dev-chain -d uniquenetwork/quartz-node-private && ./e2e/scripts/wait-first-block.sh",
+    "tests:chain-stop": "docker stop dev-chain && docker rm dev-chain",
+    "tests:prepare-scan-api": "(cd e2e/scripts ; ./prepare-scan-api.sh ../../.env)",
+    "tests:run-scan-api": "(cd e2e/scripts ; ./run-scan-api.sh ../../.env)",
     "typeorm": "node --require ts-node/register ./node_modules/typeorm/cli.js --config=common/typeorm.config.ts",
     "db:migration:generate": "npm run typeorm migration:generate -- -n",
     "db:migration:run": "npm run typeorm migration:run",
@@ -58,6 +59,7 @@
     "@types/supertest": "^2.0.11",
     "@typescript-eslint/eslint-plugin": "^5.0.0",
     "@typescript-eslint/parser": "^5.0.0",
+    "@unique-nft/types": "^0.6.0-4",
     "chai": "^4.3.6",
     "eslint": "^8.0.1",
     "eslint-config-prettier": "^8.3.0",
