--- tests/scripts/run-chain.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-ENV_FILE=${1}
-export $(grep -v '^#' $ENV_FILE | xargs)
-
-git clone -b $TESTS_CHAIN_VERSION git@gitlab.uniquenetwork.dev:qa-team/run-quartz.git ../application/chain
-docker-compose -f ../application/chain/docker-compose.yml up -d --build
